14_조요셉/1주차/requirements.txt
streamlit
google-generativeai
opencv-python
requests
dart_fss
pandas